General cleanup, bug fix, and optimization
Created resolvers folder
Fixed advanced search jumping on click
Created cache for fixed resolvers: dates, folder-tree, folder, metadataList, and modules (Code source: https://stackoverflow.com/questions/51484623/angular-6-only-call-a-resolver-once)
